+++ src/backend/Domain/Movements/IMovementStrategy.cs
@@ -6,5 +6,5 @@ namespace Domain.Movements;
 public interface IMovementStrategy
 {
     MovementType Type { get; }
-    Result Apply(Account account, decimal amount);
+    Result Apply(Account account, long amount);
 }
